pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arklint"
version = "0.1.0"
description = "The architectural rulebook for your codebase. Prevention, not detection."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["architecture", "linting", "cli", "code-quality", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "PyYAML>=6.0",
    "pathspec>=0.12.0",
]

[project.scripts]
arklint = "arklint.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
